Refactored Gabo's team changes
