+++ empleado/cocinero.py
@@ -1,0 +1,4 @@
+class Cocinero:
+    def __init__(self, salario: int, propinas: int) -> None:
+        self.salario: int = salario
+        self.propinas: int = propinas
