+++ empleado/empleado.py
@@ -1,0 +1,5 @@
+class Empleado:
+    def __init__(self, nombre:str, rfc:str, tipo:str) -> None:
+        self.nombre:str = nombre
+        self.rfc:str = rfc
+        self.tipo:str = tipo
