+++ empleado/gerente.py
@@ -1,0 +1,3 @@
+class Gerente:
+    def __init__(self, salario:int) -> None:
+        self.salario = salario
